Create Data-Pull
This script was the precursor to pipelining part data to a data lake. Instead of updating an existing database, the user selects a date range from which they would like to see part data, and the script generates a report containing only those parts.

Although it was obsoleted when the data lake was built, this script outlines the method in which the part data is organized and reported. It might have a few typos...